requirements.txt
Pillow
PyAndroidTouch
OjoPyADB